refactor: add input dir for components related input
